Initial Commit
This script will be used to receive a **POST** request with the information needed to build a **SQL** query.

The target table in the *Database* and the *Where* clause will be found and composed from the **POST** data.

Each individual table will have a different set of data to display, which will be compiled in the functions used to build the **HTML** that will be sent back (using the **card-template.html**).